Class1223

+++ src/mypackage/Homework1.java
@@ -4,7 +4,7 @@ public class Homework1 {
     public static void main(String[] args) {
         byte minimumLegalAge = 21;
         short averageMonthlySalary = 5000;
-        long miledBetweenEarthAndMoon = 238900L;
+        int miledBetweenEarthAndMoon = 238900;
         long USDebt = 23008409564495l;
         float interestRate = 3.836f;
         double itemPrice = 33.99;
